requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.1.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
chromedriver-binary==110.0.5481.30.0
colorama==0.4.6
databricks==0.2
databricks-sql-connector==2.1.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==1.2.0
frozenlist==1.3.3
h11==0.14.0
idna==3.4
ipython-genutils==0.2.0
jupyter_core==5.0.0
matplotlib-inline==0.1.6
MouseInfo==0.1.3
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.23.4
oauthlib==3.2.2
openpyxl==3.1.0
outcome==1.2.0
packaging==21.3
pandas==1.5.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.5.4
prompt-toolkit==3.0.32
psutil==5.9.4
pure-eval==0.2.2
pyarrow==9.0.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.13.0
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytweening==1.0.4
pytz==2022.6
pywin32==305
pyzmq==24.0.1
requests==2.28.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.6.1
thrift==0.13.0
tornado==6.2
tqdm==4.65.0
traitlets==5.5.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
wcwidth==0.2.5
webdriver-manager==3.8.5
wsproto==1.2.0
yarl==1.8.2
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.1.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
chromedriver-binary==110.0.5481.30.0
colorama==0.4.6
databricks==0.2
databricks-sql-connector==2.1.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==1.2.0
frozenlist==1.3.3
h11==0.14.0
idna==3.4
ipython-genutils==0.2.0
jupyter_core==5.0.0
matplotlib-inline==0.1.6
MouseInfo==0.1.3
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.23.4
oauthlib==3.2.2
openpyxl==3.1.0
outcome==1.2.0
packaging==21.3
pandas==1.5.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.5.4
prompt-toolkit==3.0.32
psutil==5.9.4
pure-eval==0.2.2
pyarrow==9.0.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.13.0
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytweening==1.0.4
pytz==2022.6
pywin32==305
pyzmq==24.0.1
requests==2.28.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.6.1
thrift==0.13.0
tornado==6.2
tqdm==4.65.0
traitlets==5.5.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
wcwidth==0.2.5
webdriver-manager==3.8.5
wsproto==1.2.0
yarl==1.8.2
